+++ tests.py
@@ -1,43 +1,46 @@
-import os
-import pykeepass
-import tempfile
+"""Test convert.py
 
+"""
+import os
+import tempfile
 import unittest
 from unittest.mock import patch
 
-import lib
+import pykeepass
+
+import convert
 
 __MASTER_PASS__ = '123456'
 
 def validate_keepass(self):
-    ## Load KeePass
-    kp = pykeepass.PyKeePass(self.output, password=__MASTER_PASS__)
+    # Load KeePass
+    kpo = pykeepass.PyKeePass(self.output, password=__MASTER_PASS__)
 
-    ## Validate folder1 group
-    folder1 = kp.find_groups(name='folder1', first=True)
+    # Validate folder1 group
+    folder1 = kpo.find_groups(name='folder1', first=True)
     self.assertIsNotNone(folder1)
     self.assertEqual(len(folder1.entries), 1)
 
-    pass1 = kp.find_entries(title='pass1', first=True, group=folder1)
+    pass1 = kpo.find_entries(title='pass1', first=True, group=folder1)
     self.assertIsNotNone(pass1)
 
     self.assertEqual(pass1.username, 'admin')
     self.assertEqual(pass1.password, '123456')
     self.assertEqual(pass1.url, 'https://localhost')
 
-    ## Validate folder2 group
-    folder2 = kp.find_groups(name='folder2', first=True)
+    # Validate folder2 group
+    folder2 = kpo.find_groups(name='folder2', first=True)
     self.assertIsNotNone(folder2)
     self.assertEqual(len(folder2.entries), 2)
 
-    pass1 = kp.find_entries(title='pass1', first=True, group=folder2)
+    pass1 = kpo.find_entries(title='pass1', first=True, group=folder2)
     self.assertIsNotNone(pass1)
 
     self.assertEqual(pass1.username, 'admin')
     self.assertEqual(pass1.password, '123457')
     self.assertEqual(pass1.url, 'https://localhost2')
 
-    pass2 = kp.find_entries(title='pass2', first=True, group=folder2)
+    pass2 = kpo.find_entries(title='pass2', first=True, group=folder2)
     self.assertIsNotNone(pass2)
 
     self.assertEqual(pass2.username, 'admin2')
@@ -48,16 +51,15 @@ class InteractiveTest(unittest.TestCase):
     def setUp(self):
         del os.environ['BITWARDEN_PASS']
 
-        (h, output) = tempfile.mkstemp()
-        self.h = h
+        _, output = tempfile.mkstemp()
         self.output = output
 
     @patch("getpass.getpass", create=True)
     def test_convert(self, getpass_func):
-        input_file = os.path.join(os.path.dirname(__file__), '..', 'resources', 'test.json')
+        input_file = os.path.join(os.path.dirname(__file__), 'resources', 'test.json')
         getpass_func.return_value = __MASTER_PASS__
 
-        lib.convert(input_file, self.output)
+        convert.convert(input_file, self.output)
 
         validate_keepass(self)
 
@@ -69,14 +71,13 @@ class NonInteractiveTest(unittest.TestCase):
     def setUp(self):
         os.environ['BITWARDEN_PASS'] = __MASTER_PASS__
 
-        (h, output) = tempfile.mkstemp()
-        self.h = h
+        _, output = tempfile.mkstemp()
         self.output = output
 
     def test_convert(self):
-        input_file = os.path.join(os.path.dirname(__file__), '..', 'resources', 'test.json')
+        input_file = os.path.join(os.path.dirname(__file__), 'resources', 'test.json')
 
-        lib.convert(input_file, self.output)
+        convert.convert(input_file, self.output)
 
         validate_keepass(self)
 
@@ -88,27 +89,26 @@ class DuplicateTest(unittest.TestCase):
     def setUp(self):
         os.environ['BITWARDEN_PASS'] = __MASTER_PASS__
 
-        (h, output) = tempfile.mkstemp()
-        self.h = h
+        _, output = tempfile.mkstemp()
         self.output = output
 
     def test_convert(self):
-        input_file = os.path.join(os.path.dirname(__file__), '..', 'resources', 'test_duplicate.json')
+        input_file = os.path.join(os.path.dirname(__file__), 'resources', 'test_duplicate.json')
 
-        lib.convert(input_file, self.output)
+        convert.convert(input_file, self.output)
 
-        ## Load KeePass
-        kp = pykeepass.PyKeePass(self.output, password=__MASTER_PASS__)
+        # Load KeePass
+        kpo = pykeepass.PyKeePass(self.output, password=__MASTER_PASS__)
 
-        ## Validate folder2 group
-        folder2 = kp.find_groups(name='folder2', first=True)
+        # Validate folder2 group
+        folder2 = kpo.find_groups(name='folder2', first=True)
         self.assertIsNotNone(folder2)
         self.assertEqual(len(folder2.entries), 2)
 
-        pass1 = kp.find_entries(title='pass1', first=True, group=folder2)
+        pass1 = kpo.find_entries(title='pass1', first=True, group=folder2)
         self.assertIsNotNone(pass1)
 
-        pass1_ = kp.find_entries(title='pass1 (1)', first=True, group=folder2)
+        pass1_ = kpo.find_entries(title='pass1 (1)', first=True, group=folder2)
         self.assertIsNotNone(pass1_)
 
         self.assertEqual(pass1_.username, 'admin')
@@ -118,6 +118,7 @@ class DuplicateTest(unittest.TestCase):
     def tearDown(self):
         if os.path.exists(self.output):
             os.unlink(self.output)
+
 
 if __name__ == 'main':
     unittest.main()
